+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **DecaLog** are documented in this *changelog*.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased - will be 1.11.0]
+### Added
+- New logger to send logs to Google Stackdriver via Google-Fluentd.
 ### Changed
 - Improved IP detection for multi-proxying.
 - Double quotes have been replaced by a grave accent in event message, context and extra.